OBJのカウント集計をlongに変更しオーバーフローを防止

+++ YukkuriMovieMaker.Plugin.Community/Shape/Model3D/Parsers/Counts.cs
@@ -2,8 +2,8 @@ namespace YukkuriMovieMaker.Plugin.Community.Shape.Model3D.Parsers;
 
 internal struct Counts
 {
-    public int V;
-    public int Vt;
-    public int Vn;
-    public int F;
+    public long V;
+    public long Vt;
+    public long Vn;
+    public long F;
 }
